+++ public/components/common/form/Select.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { Failure, classSet } from "@fider/services";
-import { ValidationContext } from "./Form2";
+import { ValidationContext } from "./Form";
 import { DisplayError, hasError } from "./DisplayError";
 
 export interface SelectOption {
